Refactor package structure and update giveaway duration in MGiveaway.java to 6 hours (from 2)
